server/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-bcrypt
flask-migrate
pymysql
python-dotenv
